# frontend/openapi.yaml
openapi: 3.0.0
info:
  title: Stark Byte Hub API
  version: 1.0.0
  description: Documentation de l’API du backend Stark Byte Hub
servers:
  - url: http://localhost:5000
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: token
...
paths:
...
  /users:
    get:
      summary: Récupère la liste de tous les utilisateurs
      tags:
        - Users
      security:
        - cookieAuth: []
      responses:
        "200":
          description: Liste des utilisateurs
    post:
      summary: Crée un nouvel utilisateur (admin uniquement)
      tags:
        - Users
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - firstName
                - lastName
                - email
                - roleId
              properties:
                firstName:
                  type: string
                lastName:
                  type: string
                email:
                  type: string
                roleId:
                  type: string
      responses:
        "201":
          description: Utilisateur créé
  /users/{id}:
    get:
      summary: Récupère un utilisateur par son ID
      tags:
        - Users
      security:
        - cookieAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Utilisateur trouvé
        "404":
          description: Utilisateur non trouvé
    patch:
      summary: Met à jour un utilisateur (admin uniquement)
      tags:
        - Users
      security:
        - cookieAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                firstName:
                  type: string
                lastName:
                  type: string
                email:
                  type: string
                roleId:
                  type: string
      responses:
        "200":
          description: Utilisateur mis à jour
    delete:
      summary: Supprime un utilisateur
      tags:
        - Users
      security:
        - cookieAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Utilisateur supprimé
  /users/me:
    patch:
      summary: Met à jour le profil de l’utilisateur connecté
      tags:
        - Users
      security:
        - cookieAuth: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                address:
                  type: string
                address2:
                  type: string
                zipCode:
                  type: string
                city:
                  type: string
                country:
                  type: string
      responses:
        "200":
          description: Profil mis à jour
  /users/password:
    patch:
      summary: Change le mot de passe de l’utilisateur connecté
      tags:
        - Users
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - currentPassword
                - newPassword
              properties:
                currentPassword:
                  type: string
                newPassword:
                  type: string
      responses:
        "200":
          description: Mot de passe mis à jour
        "401":
          description: Mot de passe actuel incorrect
...